Refactor: Move transaction related methods to TransactionValidation
