+++ core/src/main/java/bisq/core/trade/validation/TradeValidation.java
@@ -17,7 +17,6 @@
 
 package bisq.core.trade.validation;
 
-import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.support.dispute.mediation.mediator.Mediator;
 import bisq.core.trade.protocol.TradeMessage;
 import bisq.core.user.User;
@@ -29,23 +28,8 @@ import bisq.common.crypto.PubKeyRing;
 import bisq.common.crypto.Sig;
 import bisq.common.util.Base64;
 
-import org.bitcoinj.core.Address;
-import org.bitcoinj.core.AddressFormatException;
-import org.bitcoinj.core.ECKey;
-import org.bitcoinj.core.Sha256Hash;
-import org.bitcoinj.core.SignatureDecodeException;
-import org.bitcoinj.core.Transaction;
-import org.bitcoinj.core.TransactionInput;
-import org.bitcoinj.core.VerificationException;
-
-import com.google.common.annotations.VisibleForTesting;
-
 import java.security.PublicKey;
 
-import java.math.BigInteger;
-
-import java.util.Arrays;
-import java.util.Locale;
 import java.util.concurrent.TimeUnit;
 
 import lombok.extern.slf4j.Slf4j;
@@ -67,11 +51,6 @@ public final class TradeValidation {
 
     private TradeValidation() {
     }
-
-
-    /* --------------------------------------------------------------------- */
-    // Trade
-    /* --------------------------------------------------------------------- */
 
     public static String checkTradeId(String tradeId, TradeMessage tradeMessage) {
         checkArgument(isTradeIdValid(tradeId, tradeMessage), "TradeId %s is not valid", tradeId);
@@ -103,89 +82,6 @@ public final class TradeValidation {
     /* --------------------------------------------------------------------- */
     // Bitcoin
     /* --------------------------------------------------------------------- */
-
-    public static String checkBitcoinAddress(String bitcoinAddress, BtcWalletService btcWalletService) {
-        checkNonBlankString(bitcoinAddress, "bitcoinAddress");
-        checkNotNull(btcWalletService, "btcWalletService must not be null");
-
-        try {
-            Address.fromString(btcWalletService.getParams(), bitcoinAddress).getOutputScriptType();
-            return bitcoinAddress;
-        } catch (AddressFormatException | IllegalStateException e) {
-            throw new IllegalArgumentException("Invalid bitcoin address: " + bitcoinAddress, e);
-        }
-    }
-
-
-    @VisibleForTesting
-    static boolean hasSignatureData(TransactionInput input) {
-        return input.getScriptBytes().length > 0 || input.hasWitness();
-    }
-
-    public static byte[] checkDerEncodedEcdsaSignature(byte[] bitcoinSignature) {
-        checkNonEmptyBytes(bitcoinSignature, "bitcoinSignature");
-        try {
-            ECKey.ECDSASignature signature = ECKey.ECDSASignature.decodeFromDER(bitcoinSignature);
-            checkArgument(Arrays.equals(bitcoinSignature, signature.encodeToDER()),
-                    "bitcoinSignature must be strictly DER encoded");
-            checkArgument(isValidSignatureValue(signature.r),
-                    "bitcoinSignature r value is outside of allowed range");
-            checkArgument(isValidSignatureValue(signature.s),
-                    "bitcoinSignature s value is outside of allowed range");
-            checkArgument(signature.isCanonical(),
-                    "bitcoinSignature must use low-S canonical encoding");
-            return bitcoinSignature;
-        } catch (SignatureDecodeException e) {
-            throw new IllegalArgumentException("Invalid bitcoin signature", e);
-        }
-    }
-
-    @VisibleForTesting
-    static boolean isValidSignatureValue(BigInteger value) {
-        return value.signum() > 0 && value.compareTo(ECKey.CURVE.getN()) < 0;
-    }
-
-    public static Transaction checkTransaction(Transaction transaction) {
-        checkNotNull(transaction, "transaction must not be null");
-        try {
-            transaction.verify();
-        } catch (VerificationException e) {
-            throw new IllegalArgumentException("Invalid transaction", e);
-        }
-        return transaction;
-    }
-
-    public static byte[] checkSerializedTransaction(byte[] serializedTransaction,
-                                                    BtcWalletService btcWalletService) {
-        checkNonEmptyBytes(serializedTransaction, "serializedTransaction");
-        checkNotNull(btcWalletService, "btcWalletService must not be null");
-        toVerifiedTransaction(serializedTransaction, btcWalletService);
-        return serializedTransaction;
-    }
-
-    public static Transaction toVerifiedTransaction(byte[] serializedTransaction,
-                                                    BtcWalletService btcWalletService) {
-        checkNonEmptyBytes(serializedTransaction, "serializedTransaction");
-        checkNotNull(btcWalletService, "btcWalletService must not be null");
-
-        try {
-            Transaction transaction = new Transaction(btcWalletService.getParams(), serializedTransaction);
-            transaction.verify();
-            return transaction;
-        } catch (Exception e) {
-            throw new IllegalArgumentException("Invalid serialized transaction", e);
-        }
-    }
-
-    public static String checkTransactionId(String txId) {
-        checkNonBlankString(txId, "txId");
-
-        try {
-            return Sha256Hash.wrap(txId.toLowerCase(Locale.ROOT)).toString();
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("Invalid transaction ID: " + txId, e);
-        }
-    }
