+++ core/src/test/java/bisq/core/trade/validation/TradeValidationTest.java
@@ -26,24 +26,12 @@ import bisq.common.crypto.PubKeyRing;
 import bisq.common.crypto.Sig;
 import bisq.common.util.Base64;
 
-import org.bitcoinj.core.ECKey;
-import org.bitcoinj.core.SegwitAddress;
-import org.bitcoinj.core.Transaction;
-import org.bitcoinj.params.MainNetParams;
-import org.bitcoinj.params.TestNet3Params;
-
 import java.security.KeyPair;
 
 import java.nio.charset.StandardCharsets;
 
-import java.math.BigInteger;
-
-import java.util.Arrays;
-import java.util.Locale;
-
 import org.junit.jupiter.api.Test;
 
-import static bisq.core.trade.validation.TradeValidationTestUtils.btcWalletService;
 import static bisq.core.trade.validation.TradeValidationTestUtils.pubKeyRing;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -96,197 +84,6 @@ public class TradeValidationTest {
         assertThrows(NullPointerException.class, () -> TradeValidation.isTradeIdValid("trade-id", null));
     }
 
-
-    @Test
-    void checkBitcoinAddressAcceptsAddressForWalletNetwork() {
-        String bitcoinAddress = SegwitAddress.fromKey(MainNetParams.get(), new ECKey()).toString();
-
-        assertSame(bitcoinAddress, TradeValidation.checkBitcoinAddress(bitcoinAddress,
-                btcWalletService()));
-    }
-
-    @Test
-    void checkBitcoinAddressRejectsInvalidAddress() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkBitcoinAddress("not-a-bitcoin-address",
-                btcWalletService()));
-    }
-
-    @Test
-    void checkBitcoinAddressRejectsAddressFromDifferentNetwork() {
-        String testnetAddress = SegwitAddress.fromKey(TestNet3Params.get(), new ECKey()).toString();
-
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkBitcoinAddress(testnetAddress,
-                btcWalletService()));
-    }
-
-    @Test
-    void checkBitcoinAddressRejectsNullAddress() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.checkBitcoinAddress(null,
-                btcWalletService()));
-    }
-
-    @Test
-    void checkBitcoinAddressRejectsNullWalletService() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.checkBitcoinAddress(
-                SegwitAddress.fromKey(MainNetParams.get(), new ECKey()).toString(),
-                null));
-    }
-
-    @Test
-    void checkSerializedTransactionAcceptsValidSerializedTransaction() {
-        byte[] serializedTransaction = TradeValidationTestUtils.serializedTransaction();
-
-        assertSame(serializedTransaction, TradeValidation.checkSerializedTransaction(serializedTransaction,
-                btcWalletService()));
-    }
-
-    @Test
-    void checkSerializedTransactionRejectsMalformedSerializedTransaction() {
-        assertThrows(RuntimeException.class, () -> TradeValidation.checkSerializedTransaction(new byte[]{1, 2, 3},
-                btcWalletService()));
-    }
-
-    @Test
-    void checkSerializedTransactionRejectsNullSerializedTransaction() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.checkSerializedTransaction(null,
-                btcWalletService()));
-    }
-
-    @Test
-    void checkTransactionAcceptsValidTransaction() {
-        Transaction transaction = TradeValidationTestUtils.transaction(new byte[]{});
-
-        assertSame(transaction, TradeValidation.checkTransaction(transaction));
-    }
-
-    @Test
-    void checkTransactionRejectsStructurallyInvalidTransaction() {
-        Transaction transaction = TradeValidationTestUtils.transactionWithoutOutputs();
-
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkTransaction(transaction));
-    }
-
-    @Test
-    void checkTransactionRejectsNullTransaction() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.checkTransaction(null));
-    }
-
-    @Test
-    void toTransactionParsesValidSerializedVerifiedTransaction() {
-        byte[] serializedTransaction = TradeValidationTestUtils.serializedTransaction();
-
-        assertArrayEquals(serializedTransaction,
-                TradeValidation.toVerifiedTransaction(serializedTransaction, btcWalletService())
-                        .bitcoinSerialize());
-    }
-
-    @Test
-    void toVerifiedTransactionRejectsMalformedSerializedTransaction() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.toVerifiedTransaction(new byte[]{1, 2, 3},
-                btcWalletService()));
-    }
-
-    @Test
-    void toVerifiedTransactionRejectsStructurallyInvalidTransaction() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.toVerifiedTransaction(
-                TradeValidationTestUtils.serializedTransactionWithoutOutputs(),
-                btcWalletService()));
-    }
-
-    @Test
-    void toVerifiedTransactionRejectsNullSerializedTransaction() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.toVerifiedTransaction(null,
-                btcWalletService()));
-    }
-
-    @Test
-    void toVerifiedTransactionRejectsNullWalletService() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.toVerifiedTransaction(
-                TradeValidationTestUtils.serializedTransaction(),
-                null));
-    }
-
-
-
-    @Test
-    void checkTransactionIdAcceptsValidTransactionId() {
-        assertEquals(TradeValidationTestUtils.VALID_TRANSACTION_ID, TradeValidation.checkTransactionId(TradeValidationTestUtils.VALID_TRANSACTION_ID));
-    }
-
-    @Test
-    void checkTransactionIdAcceptsUpperCaseTransactionId() {
-        String transactionId = TradeValidationTestUtils.VALID_TRANSACTION_ID.toUpperCase(Locale.ROOT);
-        assertEquals(transactionId.toLowerCase(Locale.ROOT), TradeValidation.checkTransactionId(transactionId));
-    }
-
-    @Test
-    void checkTransactionIdRejectsInvalidLength() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkTransactionId(
-                TradeValidationTestUtils.VALID_TRANSACTION_ID.substring(1)));
-    }
-
-    @Test
-    void checkTransactionIdRejectsNonHexTransactionId() {
-        String transactionId = TradeValidationTestUtils.VALID_TRANSACTION_ID.substring(0, TradeValidationTestUtils.VALID_TRANSACTION_ID.length() - 1) + "g";
-
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkTransactionId(transactionId));
-    }
-
-    @Test
-    void checkTransactionIdRejectsNullTransactionId() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.checkTransactionId(null));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureAcceptsStrictDerEncodedCanonicalSignature() {
-        byte[] bitcoinSignature = TradeValidationTestUtils.bitcoinSignature(BigInteger.ONE, BigInteger.ONE);
-
-        assertSame(bitcoinSignature, TradeValidation.checkDerEncodedEcdsaSignature(bitcoinSignature));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureRejectsNullSignature() {
-        assertThrows(NullPointerException.class, () -> TradeValidation.checkDerEncodedEcdsaSignature(null));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureRejectsEmptySignature() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkDerEncodedEcdsaSignature(new byte[0]));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureRejectsMalformedSignature() {
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkDerEncodedEcdsaSignature(new byte[]{1, 2, 3}));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureRejectsNonStrictDerEncoding() {
-        byte[] bitcoinSignature = TradeValidationTestUtils.bitcoinSignature(BigInteger.ONE, BigInteger.ONE);
-        byte[] bitcoinSignatureWithTrailingData = Arrays.copyOf(bitcoinSignature, bitcoinSignature.length + 1);
-        bitcoinSignatureWithTrailingData[bitcoinSignature.length] = 1;
-
-        assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkDerEncodedEcdsaSignature(bitcoinSignatureWithTrailingData));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureRejectsValuesOutsideCurveOrder() {
-        assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkDerEncodedEcdsaSignature(TradeValidationTestUtils.bitcoinSignature(BigInteger.ZERO, BigInteger.ONE)));
-        assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkDerEncodedEcdsaSignature(TradeValidationTestUtils.bitcoinSignature(BigInteger.ONE, BigInteger.ZERO)));
-        assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkDerEncodedEcdsaSignature(TradeValidationTestUtils.bitcoinSignature(ECKey.CURVE.getN(), BigInteger.ONE)));
-        assertThrows(IllegalArgumentException.class,
-                () -> TradeValidation.checkDerEncodedEcdsaSignature(TradeValidationTestUtils.bitcoinSignature(BigInteger.ONE, ECKey.CURVE.getN())));
-    }
-
-    @Test
-    void checkDerEncodedEcdsaSignatureRejectsNonCanonicalSValue() {
-        byte[] bitcoinSignature = TradeValidationTestUtils.bitcoinSignature(BigInteger.ONE, ECKey.CURVE.getN().subtract(BigInteger.ONE));
-
-        assertThrows(IllegalArgumentException.class, () -> TradeValidation.checkDerEncodedEcdsaSignature(bitcoinSignature));
-    }
 
 
     @Test
